login y logout sin hash

+++ src/controllers/controllers.py
@@ -1,27 +1,38 @@
-from flask import request, Blueprint, jsonify, render_template, redirect, url_for, jsonify
+from flask import request, Blueprint, jsonify, render_template, redirect, url_for, jsonify, session
 from utils.db import db
-
+import hashlib
 from models.anfitrion import Anfitrion 
 from models.cliente import Cliente
 from models.reunion import Reunion 
+from models.usuario import Usuario 
 
 reuniones = Blueprint("reuniones", __name__)
+
+@reuniones.before_request
+def middleware():
+    print("Middleware")
 
 ### METODOS GET ###
 
 @reuniones.route('/reuniones')
 def getAllReuniones():
     # reuniones = db.session.execute(db.select(Reunion)).scalars()
+    if not 'loggedin' in session:
+        return redirect(url_for('reuniones.login'))
     reuniones = Reunion.query.all()
     return render_template('layout.html', reuniones=reuniones)
 
 @reuniones.route('/reunion/<int:id>')
 def getReunionById(id):
+    if not 'loggedin' in session:
+        return redirect(url_for('reuniones.login'))
     reunion = db.get_or_404(Reunion, id)
     return jsonify(Reunion.serialize(reunion))
 
 @reuniones.route('/cliente/<string:nombre>')
 def getReunionsByCliente(nombre):
+    if not 'loggedin' in session:
+        return redirect(url_for('reuniones.login'))
 
     cliente = db.session.execute(db.select(Cliente).filter_by(nombre=nombre)).fetchone()
 
@@ -69,6 +80,8 @@ def agendarReunion():
 
 @reuniones.route('/reuniones/<int:id>', methods=["GET", "PUT"])
 def editarReunion(id):
+    if not 'loggedin' in session:
+        return redirect(url_for('reuniones.login'))
     reunion = Reunion.query.get(id)
 
     anfitriones = db.session.execute(db.select(Anfitrion)).scalars()
@@ -88,8 +101,53 @@ def editarReunion(id):
 
 @reuniones.route('/<int:id>', methods=["DELETE"])
 def eliminarReunion(id):
+    if not 'loggedin' in session:
+        return redirect(url_for('reuniones.login'))
     print(id)
     reunion = Reunion.query.get(id)
     db.session.delete(reunion)
     db.session.commit()
     return "eliminado"
+
+## Login
+
+@reuniones.route('/login', methods=["GET", "POST"])
+def login():
+    if request.method == 'POST' and 'username' in request.form and 'password' in request.form:
+        # Create variables for easy access
+        username = request.form['username']
+        password = request.form['password']
+        # Retrieve the hashed password
+        # hash = password
+        # hash = hashlib.sha1(hash.encode())
+        # password = hash.hexdigest()
+
+        # Check if account exists using MySQL
+        try:
+            account = db.session.execute(db.select(Usuario).filter_by(username=username)).fetchone()[0]
+            print(account)
+        except:
+            return 'La cuenta no existe'
+         # Create session data, we can access this data in other routes
+        if account.password != password:
+            return 'La contraseña es incorrecta'
+        session['loggedin'] = True
+        session['id'] = account.id
+        print(session['id'])
+        session['username'] = account.username
+        # Redirect to home page
+        print("LOGGED")
+        return redirect(url_for('reuniones.getAllReuniones'))
+
+    if 'loggedin' in session:
+        return redirect(url_for('reuniones.getAllReuniones'))
+    return render_template("login.html")
+
+@reuniones.route('/logout')
+def logout():
+    # Remove session data, this will log the user out
+   session.pop('loggedin', None)
+   session.pop('id', None)
+   session.pop('username', None)
+   # Redirect to login page
+   return redirect(url_for('reuniones.login'))
